requirements.txt
psycopg2-binary
SpeechRecognition
pyaudio
python-dotenv
tk